+++ src/main/java/com/d2y/d2yspringapi/repositories/TokenRepository.java
@@ -8,13 +8,13 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.d2y.d2yspringapi.models.Token;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
     @Query(value = """
             SELECT t FROM Token t INNER JOIN User u\s
             ON t.user.id = u.id\s
             WHERE u.id = :id AND (t.expired = false OR t.revoked = false)\s
             """)
-    List<Token> findAllValidTokenByUser(Integer id);
+    List<Token> findAllValidTokenByUser(Long id);
 
     Optional<Token> findByToken(String token);
 }
